refactor: body, forward 관련 상태값, content-type, html버전 설정 내부 구현으로 수정
