+++ src/test/java/http/supoort/ResponseMessageConverterTest.java
@@ -1,12 +1,9 @@
 package http.supoort;
 
-import http.model.ContentType;
-import http.model.HttpProtocols;
 import http.model.HttpResponse;
 import http.model.HttpStatus;
 import org.junit.jupiter.api.Test;
 
-import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ResponseMessageConverterTest {
@@ -14,7 +11,6 @@ class ResponseMessageConverterTest {
     void NOT_FOUND_convert_테스트() {
         HttpResponse httpResponse = new HttpResponse.Builder()
                 .forward("./templates/404_ERROR.html")
-                .protocols(HttpProtocols.HTTP1_1)
                 .status(HttpStatus.NOT_FOUND)
                 .build();
 
@@ -34,9 +30,6 @@ class ResponseMessageConverterTest {
     void Forward_convert_테스트() {
         HttpResponse httpResponse = new HttpResponse.Builder()
                 .forward("./templates/index.html")
-                .protocols(HttpProtocols.HTTP1_1)
-                .status(HttpStatus.OK)
-                .addHeader(CONTENT_TYPE, ContentType.HTML.getType())
                 .build();
 
         assertThat(ResponseMessageConverter.convertHeader(httpResponse).contains("HTTP/1.1 200 OK")).isTrue();
